requirements.txt
amqp==5.2.0
aniso8601==9.0.1
asgiref==3.7.2
asyncpg==0.24.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
daphne==3.0.1
Django==4.2.7
django-celery-results==2.5.1
djangorestframework==3.14.0
Flask==3.0.0
Flask-RESTful==0.3.10
future==0.18.3
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
postgres==4.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.8.6
psycopg2-pool==1.1
PyJWT==2.8.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
responses==0.24.1
serializers==0.2.4
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
wheel==0.41.3